[Rahul/Sayan] - setBid is added to croupier which sets the bid.
What changes have you made?

* Bid is added to croupier which has value and player.
setBid receives sequence of players in an array and returns a function
to set the bid value.
The value will be "pass" or a valid bid. When value is passed, the person
won't get any more chance to bid.

Why do we need this change?

* players can tell what bid they want to set in the game.
They can pass the bid.

What can this change impact?

* Upon given the sequence of players, the players will be automatically
selected to bid next and will only receive the currect bid value.