[#11] [Sayan/Prasun] Player can see the trump card.
What changes have you made?

* Some jQuery function is added.
This will get the trump suit that is already set.
Game object is added to store data while playing is on.

Why do we need this change?

* Player need to see the trump card at a particular time.
So now player is able to see the trum card ass well as other players.

What can this change impact?

* This will help to control in the time of trump showing.