[#0] [Brindaban/Prasun]	Plyers can now be in the waiting page until all the players come
What changes have you made?

* If player clicks the 'Join game' he/she will be redirected to the waiting page.
clientHandler.js is added and tests for this is also added.
Player is assigned to teams when the come.
main.js is renamed to game.js.
teamFormation.js is renamed to team.js.
waitingPage.html is added.
cardGenerator.js is renamed to cards.js.

Why do we need this change?

* Player nedds to get a status that he/she is waiting or not after loggin in.
If a player joins the game he will be part of a team.
clientHandler is added for communication between client and game.

What can this change impact?

* Now player can see that he/she is in the waiting state.
Now player is waiting as a part of a team but not knowing which he/she belongs.