[#8] [Rahul/Prasun] Player can see whose turn is now
What changes have you made?

* Restructured the return value of getStatus in game.
Included getMyCard in player's prototype.
In gameScript modified the updateChanges function.

Why do we need this change?

* To group the status related to a player.
To show the turn of a player to all the other players.
To format the representation of the return game status gameScript is changed.

What can this change impact?

* It is very much convenient for players to see the the turn and play.