[Rahul] - Player can join game without clearing cookie
What changes have you made?

*In clientHandler.js while adding a player only checking whether there is a cookie in the browser or not.
If it has cookie then player ID will be the cookie. If the browser doesnot have a cookie it just makes a new cookie by the given name and assigns that as ID of that player.

Why do we need this change?

*Before joining to the game every time player has to remove cookie and then he can join the game. By doing this change player can join without clearing his cookie.

What can this change impact?

*Player doesn't need to clear cookie every time he joins the game.