[#15] [Brindaban/Prasun] Score board is visible to everyone at the end of the game
What changes have you made?

* New score board is added in the gamePage.html at the right corner.
In status now score is also sent.

Why do we need this change?

* At the end of the game player needs to understand who has won the game.
They need to see the score also.
Previously the score was updated but was not shown to the players.

What can this change impact?

* This change will help the player to understand the the score and the winner of the game.
Later to continue for the next game the score will help.