[Rahul] - PlayedCards in game status is an array with cards with sequence of play
What changes have you made?

*Added whoAreYou method which takes 2 player ID and tells who is the 1st player to
the 2nd player

getPlayedCards method added which gives the played cards in the sequence of play
with the relation according to the requested player's prospective.

Why do we need this change?

*To get all the cards in a sequence of play in the client side.
Cards needed to get in the sequence of play to order card over another card.

What can this change impact?

*