[Sayan/Rahul] - Closing tab of one player ends game
What changes have you made?

*Added a function to handle the closing of a tab.
When a tab is closed current game will be finished and new game will be created for new
players.

Why do we need this change?

*Someone may leave a game and won't come back. Other player's can't finish the game
without 4 players playing a game. So, it is better to let everyone know someone left and
give them chance to join a new game.

What can this change impact?

*If a player leaves a game, then other players will get informed and they will be redirected
to the homepage.